+++ R/create_analytic_cohort.R
@@ -27,14 +27,11 @@
 #'   }
 #' The function inputs `cohort`, `institution`, `stage_dx`,
 #' and `regimen_drugs` correspond to the variable
-#' names in the GENIE BPC Analytic Data Guide, available on
-#' \href{https://www.synapse.org/#!Synapse:syn21241322}{Synapse}.
+#' names in the GENIE BPC Analytic Data Guide (see README and vignettes
+#' for links to data guides).
 #'
-#' @param cohort GENIE BPC Project cancer. Must be one of "NSCLC"
-#' (non-small cell lung cancer) or "CRC" (colorectal cancer). Future cohorts
-#' will include "BrCa" (breast cancer), "PANC" (pancreatic cancer),
-#' "Prostate" (prostate cancer).
-#' @param data_synapse The list returned from pull_data_synapse().
+#' @param data_synapse The item from the nested list returned from pull_data_synapse()
+#' corresponding to the cancer cohort of interest.
 #' @param index_ca_seq Index cancer sequence. Default is 1, indicating the
 #' patient's first index cancer. The index cancer is also referred to as the
 #' BPC Project cancer in the GENIE BPC Analytic Data Guide; this is the
@@ -99,51 +96,54 @@
 #' @author Jessica Lavery
 #' @export
 #'
-#' @examples
-#' if(genieBPC:::check_synapse_login() == TRUE){
-#' # Example 1 ----------------------------------
+#' @examplesIf genieBPC::check_genie_access()
+#'
+#' # Example 1 -----------------------------------
+#' # Example using package test data
+#'
+#' create_analytic_cohort(data_synapse = genieBPC::nsclc_test_data)
+#'
+#'
+#' # Example 2 ----------------------------------
 #' # Create a cohort of all patients with stage IV NSCLC adenocarcinoma and
 #' # also return all of their corresponding cancer-directed drugs
-#' nsclc_2_1 <- pull_data_synapse("NSCLC", version = "2.1-consortium")
+#' nsclc_2_0 <- pull_data_synapse("NSCLC", version = "v2.0-public")
 #'
-#' create_analytic_cohort(cohort = "NSCLC",
-#'   data_synapse = nsclc_2_1,
+#' create_analytic_cohort(data_synapse = nsclc_2_0$NSCLC_v2.0,
 #'   stage_dx = "Stage IV",
 #'   histology = "Adenocarcinoma")
 #'
-#' # Example 2 ----------------------------------
+#' # Example 3 ----------------------------------
 #' # Create a cohort of all NSCLC patients who received Cisplatin,
 #' # Pemetrexed Disodium or Cisplatin, Etoposide as their first drug regimen
 #' # for their first index NSCLC
-#' nsclc_2_1 <- pull_data_synapse("NSCLC", version = "2.1-consortium")
 #'
-#' create_analytic_cohort(cohort = "NSCLC",
-#'     data_synapse = nsclc_2_1,
+#' nsclc_2_0 <- pull_data_synapse("NSCLC", version = "v2.0-public")
+#'
+#' create_analytic_cohort(data_synapse = nsclc_2_0$NSCLC_v2.0,
 #'     regimen_drugs = c("Cisplatin, Pemetrexed Disodium",
 #'                       "Cisplatin, Etoposide"),
 #'     regimen_order = 1,
 #'     regimen_order_type = "within cancer")
 #'
-#' # Example 3 ----------------------------------
+#' # Example 4 ----------------------------------
 #' # Create a cohort of all NSCLC patients who received Cisplatin, Pemetrexed
 #' # Disodium at any time throughout the course of treatment for their
 #' # cancer diagnosis,
 #' # but in the event that the patient received the drug multiple times,
 #' # only select the first time.
-#' nsclc_2_1 <- pull_data_synapse("NSCLC", version = "2.1-consortium")
+#' nsclc_2_0 <- pull_data_synapse("NSCLC", version = "v2.0-public")
 #'
-#' create_analytic_cohort(cohort = "NSCLC",
-#'     data_synapse = nsclc_2_1,
+#' create_analytic_cohort(data_synapse = nsclc_2_0$NSCLC_v2.0,
 #'     regimen_drugs = c("Cisplatin, Pemetrexed Disodium"),
 #'     regimen_order = 1,
 #'     regimen_order_type = "within regimen")
-#'     }
+#'
 #' @import
 #' dplyr
 #' purrr
 #' stringr
-create_analytic_cohort <- function(cohort,
-                                   data_synapse,
+create_analytic_cohort <- function(data_synapse,
                                    index_ca_seq = 1,
                                    institution,
                                    stage_dx,
@@ -155,41 +155,31 @@ create_analytic_cohort <- function(cohort,
                                    return_summary = FALSE) {
 
   # check parameters
+  # cohort object
+  if (missing(data_synapse)) {
+    stop("Specify the cohort object from the nested list returned by the
+         pull_data_synapse() function.")
+  } else if (is.null(data_synapse)) {
+    stop("The object specified for data_synapse does not exist.")
+  }
+
   # cancer cohort
-  if (length(cohort) > 1) {
+  # trying to check that the pull_data_synapse object returned is specific to the cohort
+  if (min(grepl("pt_char", paste0(names(data_synapse)))) == 1) {
     stop("Specify only one cohort at a time, even if there are multiple cohorts
          in the data_synapse object.")
   }
 
-  if (!(stringr::str_to_upper(cohort) %in% c("NSCLC", "CRC", "BRCA"))) {
-    stop("Select from available cancer cohorts:
-         NSCLC, CRC, BrCa (not case sensitive)")
-  }
+  # if (!(stringr::str_to_upper(cohort) %in% c("NSCLC", "CRC", "BRCA"))) {
+  #   stop("Select from available cancer cohorts:
+  #        NSCLC, CRC, BrCa (not case sensitive)")
+  # }
+
   #  if ( sum(!grepl("^NSCLC$", cohort)>0 , !missing(institution_temp) ,
   # !grepl(c("^DFCI$|^MSK$|^VICC$|^UHN$"), institution_temp)>0 ) >0  ){
 
-  # cohort object
-  if (missing(data_synapse)) {
-    stop("Specify the object created by pull_data_synapse() function.")
-  }
-
-  # cohort matches a cohort available in data_synapse object
-  # e.g., if cohort = NSCLC then the data_synapse object has NSCLC data
-  if (grepl(cohort, unique(data_synapse$pt_char$cohort),
-            ignore.case = TRUE) == FALSE){
-    stop("Data for the specified cohort is not available in the data_synapse
-         object specified. Check the cohort and data_synapse parameters.")
-  }
-
-  # get cohort from data_synapse object
-  # position of cohort in unique list of cohorts returned
-  # in pull_data_synapse object
-  # in the case that multiple cohorts were pulled
-  coh_position <- grep(stringr::str_to_upper(cohort),
-                       unique(word(names(data_synapse), -1, sep = "_")),
-                       ignore.case = TRUE)
-  # get that cohort name and how it is capitalized in the data_synapse object
-  cohort_temp <- unique(word(names(data_synapse), -1, sep = "_"))[coh_position]
+  # get cohort name and how it is capitalized in the data_synapse object
+  cohort_temp <- unique(word(names(data_synapse), -1, sep = "_"))
 
   # alphabetize drugs in regimen to match
   # how they are stored in variable
@@ -216,8 +206,8 @@ create_analytic_cohort <- function(cohort,
     ))
   }
   # participating institutions by cohort
-  if (sum(!missing(institution), grepl("^NSCLC$", stringr::str_to_upper(cohort))
-          > 0) > 1) {
+  if (sum(!missing(institution),
+          grepl("^NSCLC$", stringr::str_to_upper(cohort_temp)) > 0) > 1) {
     if (sum(!grepl(
       c("^DFCI$|^MSK$|^VICC$|^UHN$"),
       stringr::str_to_upper(institution)
@@ -228,7 +218,7 @@ create_analytic_cohort <- function(cohort,
   }
 
   if (sum(!missing(institution), grepl("^CRC$|^BRCA$",
-                                       stringr::str_to_upper(cohort)) > 0) > 1)
+                                       stringr::str_to_upper(cohort_temp)) > 0) > 1)
     {
     if (sum(!grepl(c("^DFCI$|^MSK$|^VICC$"), stringr::str_to_upper(institution))
     > 0) > 0) {
@@ -237,10 +227,10 @@ create_analytic_cohort <- function(cohort,
     }
   }
 
-  if (missing(institution) & stringr::str_to_upper(cohort) == "NSCLC") {
+  if (missing(institution) & stringr::str_to_upper(cohort_temp) == "NSCLC") {
     institution_temp <- c("DFCI", "MSK", "UHN", "VICC")
   } else if (missing(institution) &
-             stringr::str_to_upper(cohort) %in% c("CRC", "BRCA")) {
+             stringr::str_to_upper(cohort_temp) %in% c("CRC", "BRCA")) {
     institution_temp <- c("DFCI", "MSK", "VICC")
   } else {
     institution_temp <- stringr::str_to_upper({{ institution }})