+++ tests/testthat/test-select_unique_ngs.R
@@ -1,93 +1,112 @@
+
+
 test_that("missing data_cohort", {
+  testthat::skip_if_not(genieBPC:::check_genie_access())
+
+  nsclc_data <- pull_data_synapse("NSCLC", version = "v1.1-consortium")
+  nsclc_cohort <- create_analytic_cohort(data_synapse = nsclc_data$NSCLC_v1.1)
+  crc_data <- pull_data_synapse(c("CRC"), version = "v1.1-consortium")
+
+  # run here to avoid having to run within each test
+
+
+  objs <- list(
+    "nsclc_data" = nsclc_data,
+    "crc_data" = crc_data
+  )
+
+  list2env(objs, envir = .GlobalEnv)
+
   expect_error(select_unique_ngs())
 })
 
 test_that("non-existant oncotree code", {
-  expect_error(select_unique_ngs(data_cohort = nsclc_cohort$cohort_ca,
-                                 oncotree_code = "GENIE"))
+  expect_error(select_unique_ngs(
+    data_cohort = nsclc_cohort,
+    oncotree_code = "GENIE"
+  ))
 })
 
 test_that("min/max time", {
   # a valid value is given
-  expect_error(select_unique_ngs(data_cohort = nsclc_cohort$cohort_ca,
-                                 min_max_time = "abc"))
+  expect_error(select_unique_ngs(
+    data_cohort = nsclc_cohort,
+    min_max_time = "abc"
+  ))
 
   # only 1 value is given
-  expect_error(select_unique_ngs(data_cohort = nsclc_cohort$cohort_ca,
-                                 min_max_time = c("min", "max")))
+  expect_error(select_unique_ngs(
+    data_cohort = nsclc_cohort,
+    min_max_time = c("min", "max")
+  ))
 })
 
 test_that("sample_type", {
   # a valid value is given
-  expect_error(select_unique_ngs(data_cohort = nsclc_cohort$cohort_ca,
-                                 sample_type = "abc"))
+  expect_error(select_unique_ngs(
+    data_cohort = nsclc_cohort,
+    sample_type = "abc"
+  ))
 
   # only 1 value is given
-  expect_error(select_unique_ngs(data_cohort = nsclc_cohort$cohort_ca,
-                                 sample_type = c("primary", "local")))
+  expect_error(select_unique_ngs(
+    data_cohort = nsclc_cohort,
+    sample_type = c("primary", "local")
+  ))
 })
 
 test_that("function returns unique sample for each record", {
-  # exit if user doesn't have synapser, a log in, or access to data.
-  skip_if_not_installed("synapser", minimum_version = NULL)
-  skip_if(inherits(try(synapser::synLogin(), silent = TRUE), "try-error"),
-          "Not logged into Synapse")
-  skip_if(inherits(try(synapser::synGet("syn26948075"), silent = TRUE), "try-error"),
-          "Not able to access the data")
+  skip_if_not(genieBPC:::check_genie_access())
 
-  # run here to avoid having to run within each test
-  nsclc_data <- pull_data_synapse("NSCLC", version = "1.1-consortium")
-  crc_data <- pull_data_synapse(c("CRC"), version = "1.1-consortium")
-
-  objs <- list("nsclc_data" = nsclc_data,
-               "crc_data" = crc_data)
-
-  list2env(objs, envir = .GlobalEnv)
 
   # NSCLC #
   ### all samples ###
   cohort_temp <- create_analytic_cohort(
-    cohort = "NSCLC",
-    data_synapse = nsclc_data,
+    data_synapse = nsclc_data$NSCLC_v1.1,
     return_summary = FALSE
   )
 
-  expect_warning(test1 <- select_unique_ngs(
-    data_cohort = cohort_temp$cohort_ngs))
+  test1 <- select_unique_ngs(
+    data_cohort = cohort_temp$cohort_ngs
+  )
   expect_true(tibble::is_tibble(test1))
   expect_equal(ncol(test1), 20)
   expect_equal(nrow(test1), 1849)
-  expect_equal(length(unique(test1$cpt_genie_sample_id)),
-               length(unique(test1$record_id)))
+  expect_equal(
+    length(unique(test1$cpt_genie_sample_id)),
+    length(unique(test1$record_id))
+  )
 
 
   ### Stage IV ###
   cohort_temp <- create_analytic_cohort(
-    cohort = "NSCLC",
     stage_dx = c("Stage IV"),
-    data_synapse = nsclc_data,
+    data_synapse = nsclc_data$NSCLC_v1.1,
     return_summary = FALSE
   )
 
 
-  expect_warning(test2 <- select_unique_ngs(
-    data_cohort = cohort_temp$cohort_ngs))
+  test2 <- select_unique_ngs(
+    data_cohort = cohort_temp$cohort_ngs
+  )
   expect_true(tibble::is_tibble(test2))
   expect_equal(ncol(test2), 20)
   expect_equal(nrow(test2), 793)
-  expect_equal(length(unique(test2$cpt_genie_sample_id)),
-               length(unique(test2$record_id)))
+  expect_equal(
+    length(unique(test2$cpt_genie_sample_id)),
+    length(unique(test2$record_id))
+  )
 
   ### DFCI only ###
   cohort_temp <- create_analytic_cohort(
-    cohort = "NSCLC",
-    data_synapse = nsclc_data,
+    data_synapse = nsclc_data$NSCLC_v1.1,
     return_summary = FALSE,
     institution = "DFCI"
   )
 
-  expect_warning(test3 <- select_unique_ngs(
-    data_cohort = cohort_temp$cohort_ngs))
+  test3 <- select_unique_ngs(
+    data_cohort = cohort_temp$cohort_ngs
+  )
   expect_true(tibble::is_tibble(test3))
   expect_equal(ncol(test3), 20)
   expect_equal(nrow(test3), 699)
@@ -96,12 +115,12 @@ test_that("function returns unique sample for each record", {
 
   ### Check patient ###
   ##### Local min #####
-  expect_warning(test4 <- select_unique_ngs(
+  test4 <- select_unique_ngs(
     data_cohort = cohort_temp$cohort_ngs,
     oncotree_code = "LUAD",
     sample_type = "Local",
     min_max_time = "min"
-  ))
+  )
   expect_true(tibble::is_tibble(test4))
   expect_equal(ncol(test4), 20)
   expect_equal(nrow(test4), 699)
@@ -115,12 +134,12 @@ test_that("function returns unique sample for each record", {
   )
 
   ##### Local max #####
-  expect_warning(test5 <- select_unique_ngs(
+  test5 <- select_unique_ngs(
     data_cohort = cohort_temp$cohort_ngs,
     oncotree_code = "LUAD",
     sample_type = "Local",
     min_max_time = "max"
-  ))
+  )
   expect_true(tibble::is_tibble(test5))
   expect_equal(ncol(test5), 20)
   expect_equal(nrow(test5), 699)
@@ -135,12 +154,12 @@ test_that("function returns unique sample for each record", {
 
 
   ##### Met min/max (are the same) #####
-  expect_warning(test6 <- select_unique_ngs(
+  test6 <- select_unique_ngs(
     data_cohort = cohort_temp$cohort_ngs,
     oncotree_code = "LUAD",
     sample_type = "Metastasis",
     min_max_time = "max"
-  ))
+  )
   expect_true(tibble::is_tibble(test6))
   expect_equal(ncol(test6), 20)
   expect_equal(nrow(test6), 699)