+++ NEWS.md
@@ -1,3 +1,13 @@
+# genieBPC (development version)
+
+* Removed TramineR from dependencies file (issue # 50)
+
+* Pull the most recent public and consortium data (issue #42)
+
+* Removed synapser from dependencies file (issue #45)
+
+* Fixed `cpt_sample_type`/`sample_type` (issue #43)
+
 # genieBPC 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.